+++ src/client/tabbing.ts
@@ -5,7 +5,7 @@ import {Identity} from 'hadouken-js-adapter';
 
 import {tryServiceDispatch} from './connection';
 import {AddTabPayload, getId, parseIdentity, SetTabstripPayload, TabAPI, UpdateTabPropertiesPayload} from './internal';
-import {ApplicationUIConfig, TabAddedPayload, TabGroupEventPayload, TabProperties, TabPropertiesUpdatedPayload, WindowIdentity, TabGroupMaximizedPayload, TabGroupRestoredPayload, TabGroupMinimizedPayload} from './types';
+import {ApplicationUIConfig, TabAddedPayload, TabGroupEventPayload, TabProperties, TabPropertiesUpdatedPayload, WindowIdentity} from './types';
 
 /**
  * Fired when a window has had its tab properties updated.  See {@link addEventListener}.
@@ -36,27 +36,6 @@ export interface TabAddedEvent extends CustomEvent<TabAddedPayload> {
 }
 
 /**
- * Fired when a tab group is restored.  See {@link addEventListener}.
- */
-export interface TabGroupRestoredEvent extends CustomEvent<TabGroupRestoredPayload> {
-    type: 'tab-group-restored';
-}
-
-/**
- * Fired when a tab group is minimized.  See {@link addEventListener}.
- */
-export interface TabGroupMinimizedEvent extends CustomEvent<TabGroupMinimizedPayload> {
-    type: 'tab-group-minimized';
-}
-
-/**
- * Fired when a tab group is maximized.  See {@link addEventListener}.
- */
-export interface TabGroupMaximizedEvent extends CustomEvent<TabGroupMaximizedPayload> {
-    type: 'tab-group-maximized';
-}
-
-/**
  * @hidden
  */
 export interface EventMap {
@@ -64,9 +43,6 @@ export interface EventMap {
     'tab-removed': TabRemovedEvent;
     'tab-activated': TabActivatedEvent;
     'tab-properties-updated': TabPropertiesUpdatedEvent;
-    'tab-group-restored': TabGroupRestoredEvent;
-    'tab-group-minimized': TabGroupMinimizedEvent;
-    'tab-group-maximized': TabGroupMaximizedEvent;
 }
 
 /**
@@ -148,57 +124,6 @@ export async function addEventListener(eventType: 'tab-activated', listener: (ev
  */
 export async function addEventListener(eventType: 'tab-properties-updated', listener: (event: TabPropertiesUpdatedEvent) => void): Promise<void>;
 
-/**
- * Event fired whenever the current tab group is restored from being maximized or minimized.
- *
- * ```ts
- * import {tabbing} from 'openfin-layouts';
- *
- * tabbing.addEventListener('tab-group-restored', (event: TabGroupRestoredEvent) => {
- *     const tabGroupID = event.detail.identity;
- *     console.log(`Tab group restored: ${tabID.uuid}/${tabID.name}`);
- * });
- * ```
- *
- * @type tab-group-restored
- * @event
- */
-export async function addEventListener(eventType: 'tab-group-restored', listener: (event: TabGroupRestoredEvent) => void): Promise<void>;
-
-/**
- * Event fired whenever the current tab group is minimized.
- *
- * ```ts
- * import {tabbing} from 'openfin-layouts';
- *
- * tabbing.addEventListener('tab-group-minimized', (event: TabGroupMinimizedEvent) => {
- *     const tabGroupID = event.detail.identity;
- *     console.log(`Tab group minimized: ${tabID.uuid}/${tabID.name}`);
- * });
- * ```
- *
- * @type tab-group-minimized
- * @event
- */
-export async function addEventListener(eventType: 'tab-group-minimized', listener: (event: TabGroupMinimizedEvent) => void): Promise<void>;
-
-/**
- * Event fired whenever the current tab group is maximized.
- *
- * ```ts
- * import {tabbing} from 'openfin-layouts';
- *
- * tabbing.addEventListener('tab-group-maximized', (event: TabGroupMaximizedEvent) => {
- *     const tabGroupID = event.detail.identity;
- *     console.log(`Tab group maximized: ${tabID.uuid}/${tabID.name}`);
- * });
- * ```
- *
- * @type tab-group-minimized
- * @event
- */
-export async function addEventListener(eventType: 'tab-group-maximized', listener: (event: TabGroupMaximizedEvent) => void): Promise<void>;
-
 export async function addEventListener<K extends keyof EventMap>(eventType: K, listener: (event: EventMap[K]) => void): Promise<void> {
     if (typeof fin === 'undefined') {
         throw new Error('fin is not defined. The openfin-layouts module is only intended for use in an OpenFin application.');
@@ -207,8 +132,6 @@ export async function addEventListener<K extends keyof EventMap>(eventType: K, l
     // Without this we would need to handle multiple registration ourselves.
     window.addEventListener(eventType, listener as EventListener);
 }
-
-
 
 /**
  * Returns array of window identity references for tabs belonging to the tab group of the provided window context.
